refactor: LeaseDiagnosisController 비지니스 로직을 LeaseDiagnosisService로 분리
